[BE] refactor: 불필요 import 제거
