+++ backend/src/main/java/site/pathos/domain/annotationHistory/service/AnnotationHistoryService.java
@@ -3,14 +3,13 @@ package site.pathos.domain.annotationHistory.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import site.pathos.domain.annotation.cellAnnotation.CellAnnotationRepository;
+import site.pathos.domain.annotation.cellAnnotation.repository.CellAnnotationRepository;
 import site.pathos.domain.annotation.cellAnnotation.dto.CellDetail;
 import site.pathos.domain.annotation.cellAnnotation.entity.CellAnnotation;
 import site.pathos.domain.annotation.tissueAnnotation.entity.TissueAnnotation;
 import site.pathos.domain.annotationHistory.dto.response.AnnotationHistoryResponseDto;
 import site.pathos.domain.annotationHistory.entity.AnnotationHistory;
 import site.pathos.domain.annotationHistory.repository.AnnotationHistoryRepository;
-import site.pathos.domain.model.Repository.ModelRepository;
 import site.pathos.domain.roi.dto.request.RoiDetail;
 import site.pathos.domain.roi.dto.request.RoiPayload;
 import site.pathos.domain.roi.entity.Roi;
